refactor: classified types and constants
